requirements.txt
python-telegram-bot==13.14
requests
jsonpickle
git+https://github.com/Nikita2305/python-telegram-simple.git#egg=python-telegram-simple
